Move SPIRV-Tools wrapper functions into their own files.
